requirements.txt
langchain
langchain-community
langchainhub
transformers
sentence-transformers
faiss-cpu
openai
huggingface-hub
tqdm
python-dotenv
pypdf